refactoring: EnabledQuestionInputArea 에서 UserInfoInput 컴포넌트로 파일 분리
